Move ActiveMerchant.deprecated to its own file
This is to allow requiring files using this method
without loading all of ActiveMerchant.
