+++ lib/active_merchant.rb
@@ -51,16 +51,6 @@ require 'active_merchant/posts_data'
 require 'active_merchant/billing'
 require 'active_merchant/version'
 require 'active_merchant/country'
-
-module ActiveMerchant
-  def self.deprecated(message, caller=Kernel.caller[1])
-    warning = caller + ": " + message
-    if(respond_to?(:logger) && logger.present?)
-      logger.warn(warning)
-    else
-      warn(warning)
-    end
-  end
-end
+require 'active_merchant/deprecated'
 
 I18n.enforce_available_locales = false
